Redesign: compiling
First step of a redesign. I've removed 2 files and moved their functionality around.
Things are compiling but the redesign is not yet finished. Still some cleanup work to do. The unbundling is untested so use at your own risk. A stable version of the unbundler exists as a separate repo.